+++ NEWS.md
@@ -1,5 +1,7 @@
 # waldo (development version)
 
+* Handle S4 objects that have attributes that are not slots.
+
 # waldo 0.2.1
 
 * `compare()` now labels output as `old` and `new`, since that's the most